refactor : tokenController에서 HTTP 관련 처리하게 수정
