+++ src/main/java/corecord/dev/domain/auth/application/TokenService.java
@@ -1,8 +1,5 @@
 package corecord.dev.domain.auth.application;
 
-import corecord.dev.common.exception.GeneralException;
-import corecord.dev.common.status.ErrorStatus;
-import corecord.dev.common.util.CookieUtil;
 import corecord.dev.domain.auth.jwt.JwtUtil;
 import corecord.dev.domain.auth.domain.entity.RefreshToken;
 import corecord.dev.domain.auth.domain.entity.TmpToken;
@@ -10,17 +7,13 @@ import corecord.dev.domain.auth.status.TokenErrorStatus;
 import corecord.dev.domain.auth.exception.TokenException;
 import corecord.dev.domain.auth.domain.repository.RefreshTokenRepository;
 import corecord.dev.domain.auth.domain.repository.TmpTokenRepository;
+import corecord.dev.domain.user.application.UserDbService;
 import corecord.dev.domain.user.domain.converter.UserConverter;
 import corecord.dev.domain.user.domain.dto.response.UserResponse;
 import corecord.dev.domain.user.domain.entity.User;
-import corecord.dev.domain.user.domain.repository.UserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseCookie;
 import org.springframework.stereotype.Service;
 
 @Slf4j
@@ -29,25 +22,17 @@ import org.springframework.stereotype.Service;
 public class TokenService {
 
     private final RefreshTokenRepository refreshTokenRepository;
-    private final UserRepository userRepository;
     private final JwtUtil jwtUtil;
-    private final CookieUtil cookieUtil;
+    private final UserDbService userDbService;
     private final TmpTokenRepository tmpTokenRepository;
-
-    @Value("${jwt.access-token.expiration-time}")
-    private long accessTokenExpirationTime;
-
-    @Value("${jwt.refresh-token.expiration-time}")
-    private long refreshTokenExpirationTime;
 
     /**
      * 임시 토큰을 이용하여 AccessToken과 RefreshToken을 발급한다.
-     * @param response
      * @param tmpToken
      * @return
      */
     @Transactional
-    public UserResponse.UserDto issueTokens(HttpServletResponse response, String tmpToken) {
+    public UserResponse.UserDto issueTokens(String tmpToken) {
         // 임시 토큰 유효성 검증
         TmpToken tmpTokenEntity = validateTmpToken(tmpToken);
         tmpTokenRepository.delete(tmpTokenEntity);
@@ -60,51 +45,23 @@ public class TokenService {
 
         // 새 AccessToken 발급
         String accessToken = jwtUtil.generateAccessToken(userId);
-        log.info("AccessToken: {}", accessToken);
 
-        // 쿠키에 AccessToken 및 RefreshToken 추가
-        setTokenCookies(response, "refreshToken", refreshToken);
-        setTokenCookies(response, "accessToken", accessToken);
-
-        User user = findUserById(userId);
-        return UserConverter.toUserDto(user);
+        User user = userDbService.findUserById(userId);
+        return UserConverter.toUserDto(user, accessToken, refreshToken);
     }
 
     /**
      * RefreshToken을 이용하여 새로운 AccessToken을 발급한다.
-     * @param request
-     * @param response
+     * @param refreshToken
      * @return
      */
     @Transactional
-    public UserResponse.UserDto reissueAccessToken(HttpServletRequest request, HttpServletResponse response) {
-        String refreshToken = getRefreshTokenFromCookie(request);
-
-        // RefreshToken 유효성 검증
+    public String reissueAccessToken(String refreshToken) {
         validateRefreshToken(refreshToken);
 
-        // 새 AccessToken 발급
         Long userId = Long.parseLong(jwtUtil.getUserIdFromRefreshToken(refreshToken));
-        String newAccessToken = jwtUtil.generateAccessToken(userId);
 
-        // 기존 AccessToken 삭제
-        ResponseCookie accessTokenCookie = cookieUtil.deleteCookie("accessToken");
-        response.addHeader("Set-Cookie", accessTokenCookie.toString());
-
-        // 쿠키에 새 AccessToken 추가
-        setTokenCookies(response, "accessToken", newAccessToken);
-        log.info("AccessToken: {}", newAccessToken);
-
-        User user = findUserById(userId);
-        return UserConverter.toUserDto(user);
-    }
-
-    private String getRefreshTokenFromCookie(HttpServletRequest request) {
-        String refreshToken = cookieUtil.getCookieValue(request, "refreshToken");
-        if (refreshToken == null) {
-            throw new TokenException(TokenErrorStatus.REFRESH_TOKEN_NOT_FOUND);
-        }
-        return refreshToken;
+        return jwtUtil.generateAccessToken(userId);
     }
 
     private void validateRefreshToken(String refreshToken) {
@@ -114,7 +71,6 @@ public class TokenService {
 
         refreshTokenRepository.findByRefreshToken(refreshToken)
                 .orElseThrow(() -> new TokenException(TokenErrorStatus.REFRESH_TOKEN_NOT_FOUND));
-
     }
 
     private TmpToken validateTmpToken(String tmpToken) {
@@ -123,21 +79,5 @@ public class TokenService {
         }
         return tmpTokenRepository.findByTmpToken(tmpToken)
                 .orElseThrow(() -> new TokenException(TokenErrorStatus.TMP_TOKEN_NOT_FOUND));
-    }
-
-    // 토큰 쿠키 설정
-    private void setTokenCookies(HttpServletResponse response, String tokenName, String token) {
-        if(tokenName.equals("accessToken")) {
-            ResponseCookie accessTokenCookie = cookieUtil.createTokenCookie(tokenName, token, accessTokenExpirationTime);
-            response.addHeader("Set-Cookie", accessTokenCookie.toString());
-        } else {
-            ResponseCookie refreshTokenCookie = cookieUtil.createTokenCookie(tokenName, token, refreshTokenExpirationTime);
-            response.addHeader("Set-Cookie", refreshTokenCookie.toString());
-        }
-    }
-
-    private User findUserById(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.UNAUTHORIZED));
     }
 }
